todoListFastapi/requirements.txt
fastapi==0.75.2
uvicorn==0.17.6
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
h11==0.12.0
passlib==1.7.4
pycparser==2.20
pydantic==1.9.0
PyJWT==2.0.0
six==1.15.0
starlette==0.17.1
python-dotenv==0.20.0
motor==2.5.1
python-jose==3.3.0
aioredis==2.0.1
asyncio==3.4.3
jwt==1.3.1
pymongo==3.12.3
python-multipart==0.0.5
dnspython==2.2.1